Add tuned deadzoning with two-level biasing and separate end-of-block biasing
The idea is based on what the Thor codebase does (which is
undocumented).

See comments for details on how the constants were derived,
and the algorithm in general.

The intent is that this is used as the fast path once we add
trellis quantization.

This particulrily helps high motion clips like Red Kayak and
Ducks Take Off:

FILL ME IN.

AWCY Results:

FILL ME IN.

AWCY Link: FILL ME IN

Implements #1338 and #1340.

Signed-off-by: Derek Buitenhuis <derek.buitenhuis@gmail.com>